ai-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.60.0
httpx==0.28.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pydantic==2.10.0
python-multipart==0.0.18
